商品页

+++ database/migrations/2019_03_17_004005_add_thing_to_goods_table.php
@@ -14,9 +14,9 @@ class AddThingToGoodsTable extends Migration
     public function up()
     {
         Schema::table('goods', function (Blueprint $table) {
-            $table->smallInteger('province_code')->comment('省');
-            $table->smallInteger('city_code')->comment('市');
-            $table->smallInteger('area_code')->comment('区');
+            $table->Integer('province_code')->comment('省');
+            $table->Integer('city_code')->comment('市');
+            $table->Integer('area_code')->comment('区');
             $table->string('province_name', 20)->after('province_code')->comment('省名');
             $table->string('city_name', 20)->after('city_code')->comment('市名');
             $table->string('area_name', 20)->after('area_code')->comment('区名');
